requirements.txt
Django==1.8.5
Pillow==3.0.0
Unidecode==0.04.18
argparse==1.2.1
django-autoslug==1.9.3
django-cities-light==3.1.2
django-formset-js==0.4.3
django-jquery-js==2.1.4
django-pure-pagination==0.2.1
djangorestframework==3.2.3
psycopg2==2.6.1
python-dateutil==2.4.2
requests==2.8.1
six==1.10.0
wsgiref==0.1.2
dj-database-url==0.4.0
gunicorn==19.4.5
whitenoise==2.0.6

